Conflicts:
	tests/sysinfo-tester/main.cpp

+++ src/multimedia/effects/qsoundeffect.cpp
@@ -96,7 +96,7 @@ QT_BEGIN_NAMESPACE
 */
 
 /*!
-    \qmlproperty int SoundEffect::volume
+    \qmlproperty qreal SoundEffect::volume
 
     This property holds the volume of the playback, from 0.0 (silent) to 1.0 (maximum volume).
 */
